requirements.txt
paho-mqtt==1.5.0
pkg-resources==0.0.0
python-dotenv==0.14.0
websockets==8.1
